refactor: move logging functions to separate directory
