+++ src/logging/show-current-directory.js
